Update cerner constants location
