+++ src/applications/static-pages/health-care-manage-benefits/components/CernerCallToAction/index.js
@@ -8,8 +8,7 @@ import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 import LoadingIndicator from '@department-of-veterans-affairs/formation-react/LoadingIndicator';
 import environment from 'platform/utilities/environment';
 import { apiRequest } from 'platform/utilities/api';
-import { getCernerURL } from 'platform/utilities/constants';
-import { CERNER_FACILITY_IDS } from '../../constants';
+import { CERNER_FACILITY_IDS, getCernerURL } from 'platform/utilities/cerner';
 
 export class CernerCallToAction extends Component {
   static propTypes = {
