--- src/applications/static-pages/health-care-manage-benefits/constants.js
@@ -1,1 +1,0 @@
-export const CERNER_FACILITY_IDS = ['668', '757'];
