+++ src/platform/utilities/constants.js
@@ -1,18 +1,7 @@
-// Relative imports.
-import environment from 'platform/utilities/environment';
-
 // TODO: Rename this to something that makes more sense
 export const requestStates = {
   notCalled: 'not called',
   pending: 'pending',
   succeeded: 'succeeded',
   failed: 'failed',
-};
-
-export const getCernerURL = path => {
-  const root = environment.isProduction()
-    ? 'https://patientportal.myhealth.va.gov'
-    : 'https://ehrm-va-test.patientportal.us.healtheintent.com';
-
-  return `${root}${path}`;
 };
